+++ src/utils/VirtruWrapper.js
@@ -21,9 +21,20 @@ function _pushAction(action) {
  *
  * @param {?object} opts
  */
-function policyBuilder(opts) {
-  const builder = new Virtru.PolicyBuilder(opts);
-  let actions = [`const policy = new Virtru.PolicyBuilder(${opts ? 'policy' : ''})`];
+function policyBuilder(existingPolicy) {
+  if (existingPolicy) {
+    _pushAction({
+      title: 'Get Policy Builder',
+      code: 'const builder = existingPolicy.builder();',
+    });
+  } else {
+    _pushAction({
+      title: 'Create Policy Builder',
+      code: 'const builder = new Virtru.PolicyBuilder();',
+    });
+  }
+  const builder = existingPolicy ? existingPolicy.builder() : new Virtru.PolicyBuilder();
+  let actions = ['const policy = builder'];
   // This proxy records all calls, then logs them to the UI on `build` invocations.
   return new Proxy(builder, {
     get(target, propKey, receiver) {