+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased](https://github.com/virtru/protect-and-track/compare/master...HEAD)
 
 - NO-REF: _patch_
+  - Fix policy building by using `policy.builder()`
+- NO-REF: _patch_
   - Use Audit from SDK
 
 ## [v1.0.0](https://github.com/virtru/protect-and-track/compare/v0.1.2...v1.0.0) - 2019-08-02